# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: API de gestión de incidencias
  version: 0.1.0
  description: >-
    Contrato inicial de la plataforma. Las operaciones se implementarán por
    iteración; publicar este archivo no implica que los endpoints estén activos.
...
servers:
  - url: http://localhost:3000/api/v1
    description: Desarrollo local (desde Iteración 1)
...
paths:
...
  /incidents:
    post:
      tags: [Incidents]
      summary: Crear reporte ciudadano o preliminar de invitado
      description: >-
        Un invitado genera un reporte no verificado. El servidor fija source=MOBILE_APP.
        Idempotency-Key o clientRequestId evita duplicados; un reintento idéntico devuelve 200.
      operationId: createIncident
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncidentRequest'
      responses:
        '201':
          description: Reporte recibido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentReceipt'
        '200':
          description: Reintento idéntico con la misma referencia
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IncidentReceipt' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /incidents/mine:
    get:
      tags: [Incidents]
      summary: Listar reportes propios
      operationId: listMyIncidents
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0 } }
      responses:
        '200':
          description: Reportes del ciudadano autenticado
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IncidentSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /incidents/{id}:
    get:
      tags: [Incidents]
      summary: Obtener detalle autorizado
      operationId: getIncident
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      responses:
        '200':
          description: Detalle del reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentDetail'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /ops/incidents/phone:
    post:
      tags: [Operations]
      summary: Registrar reporte telefónico con fuente PHONE fijada por el servidor
      operationId: createPhoneIncident
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PhoneIncidentRequest' }
      responses:
        '201':
          description: Llamada registrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IncidentReceipt' }
        '200':
          description: Reintento idéntico
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IncidentReceipt' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...
  /ops/incidents:
    get:
      tags: [Operations]
      summary: Bandeja institucional filtrada por ámbito autorizado
      operationId: listOperationalIncidents
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/IncidentStatus' }
      responses:
        '200':
          description: Incidencias autorizadas
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IncidentSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IncidentId:
      name: id
      in: path
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Identificador único del intento de creación generado por el cliente.
      schema: { type: string, minLength: 1, maxLength: 128 }
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Autenticación requerida o inválida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Acceso fuera del rol o ámbito autorizado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Conflicto de estado o duplicidad
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    TooManyRequests:
      description: Límite de solicitudes superado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    UnprocessableEntity:
      description: Contraseña fuera de la política de seguridad
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
...
    PhoneIncidentRequest:
      type: object
      required: [institutionId, siteId, categoryId, description]
      properties:
        institutionId: { type: string, format: uuid }
        siteId: { type: string, format: uuid }
        categoryId: { type: integer }
        subcategoryId: { type: integer }
        description: { type: string, minLength: 1, maxLength: 2000 }
        location: { allOf: [{ $ref: '#/components/schemas/Coordinates' }], nullable: true }
        callerContact: { type: string, maxLength: 100 }
        clientRequestId: { type: string, maxLength: 128 }
...
    ApiError:
      type: object
      required: [code, message, correlationId]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string }
        correlationId: { type: string }
...
    Coordinates:
      type: object
      required: [latitude, longitude]
      properties:
        latitude: { type: number, format: double, minimum: -90, maximum: 90 }
        longitude: { type: number, format: double, minimum: -180, maximum: 180 }
        accuracyMeters: { type: number, minimum: 0 }
        capturedAt: { type: string, format: date-time }
        reference: { type: string, maxLength: 500 }
    CreateIncidentRequest:
      type: object
      required: [categoryId, location, description]
      properties:
        categoryId: { type: integer }
        subcategoryId: { type: integer }
        location: { $ref: '#/components/schemas/Coordinates' }
        description: { type: string, minLength: 1, maxLength: 2000 }
        affectedPeople: { type: integer, minimum: 0 }
        occurredAt: { type: string, format: date-time }
        clientRequestId: { type: string, maxLength: 128 }
    IncidentStatus:
      type: string
      enum: [reported, verifying, assigned, en_route, attending, resolved, closed]
    IncidentReceipt:
      type: object
      required: [id, reference, status, verified, source, createdAt]
      properties:
        id: { type: string }
        reference: { type: string }
        status: { $ref: '#/components/schemas/IncidentStatus' }
        verified: { type: boolean }
        source: { type: string, enum: [MOBILE_APP, PHONE] }
        createdAt: { type: string, format: date-time }
    IncidentSummary:
      allOf:
        - $ref: '#/components/schemas/IncidentReceipt'
        - type: object
          properties:
            categoryId: { type: integer }
            subcategoryId: { type: integer, nullable: true }
    IncidentDetail:
      allOf:
        - $ref: '#/components/schemas/IncidentSummary'
        - type: object
          properties:
            description: { type: string }
            location: { allOf: [{ $ref: '#/components/schemas/Coordinates' }], nullable: true }
            evidence: { type: array, items: { $ref: '#/components/schemas/EvidenceMetadata' } }
    EvidenceMetadata:
      type: object
      required: [id, mediaType, uploadedAt]
      properties:
        id: { type: string }
        mediaType: { type: string, enum: [image/jpeg, image/png, image/webp] }
        uploadedAt: { type: string, format: date-time }
...